requirements.txt
flask
flask-cors
flask-session
python-dotenv
pypdf2
requests
gunicorn